Refactor: Stac API code
