--- src/main/java/iudx/catalogue/server/authenticator/authorization/AuthorizationStratergy.java
@@ -1,6 +1,0 @@
-package iudx.catalogue.server.authenticator.authorization;
-
-public interface AuthorizationStratergy {
-
-  boolean isAuthorized(AuthorizationRequest authRequest);
-}
