--- src/main/java/iudx/catalogue/server/authenticator/authorization/JwtAuthorization.java
@@ -1,15 +1,0 @@
-package iudx.catalogue.server.authenticator.authorization;
-
-
-public class JwtAuthorization {
-
-  private final AuthorizationStratergy authStrategy;
-
-  public JwtAuthorization(final AuthorizationStratergy authStrategy) {
-    this.authStrategy = authStrategy;
-  }
-
-  public boolean isAuthorized(AuthorizationRequest authRequest) {
-    return authStrategy.isAuthorized(authRequest);
-  }
-}
