--- src/main/java/iudx/catalogue/server/authenticator/authorization/Method.java
@@ -1,9 +1,0 @@
-package iudx.catalogue.server.authenticator.authorization;
-
-public enum Method {
-  GET,
-  POST,
-  PUT,
-  DELETE,
-  PATCH
-}
